[BREAKING] Deprecated fetch-jsonp
We have been using jsonp to handle http requests on the browser,
and jsonp was the only way to by pass our origin policy, but it
wasn't really working for us, and we decided to deprecated it,
which will impact anyone using our rest endpoint on the browser.

We have only enable `Access-Control-Allow-Origin` on our public
endpoint (.ticker, .trades and .orderbook)

For now, the trade endpoint won't work on the browser, and you
should handle that on the server side instead, this might change
in the future.